+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+- Make ClassNameData public
+- Replace pedantic with lints
+- Add naming strategy option namely - snake, camel
+- Update README with new/deleted flags
+- Fix iconsMap hardcoded `ui_icons.json` which prevented custom out-font name
+
 ## 3.0.0
 
 * Replaced [icon-font-generator] with [fantasticon] [pull#14](https://github.com/rbcprolabs/icon_font_generator/pull/14)